SAMZA-2584: Refactor ClusterBasedJobCoordinator
Issues: In the deployment flow of a beam job, we will have a complicate flow: ClusterBasedJobCoordinator#main -> Beam main class -> JobCoordinatorLaunchUtil -> ClusterBasedJobCoordinator.
Changes:
  1. Move ClusterBasedJobCoordinator#main to ClusterBasedJobCoordinatorRunner#main
  2. Update run-jc.sh to invoke ClusterBasedJobCoordinatorRunner
Tests:
  1. unit tests
  2. Deployed hello samza job successfully with the change following instructions on http://samza.apache.org/startup/hello-samza/latest/
API Changes: None
Upgrade Instructions: None
Usage Instructions: None
